refactor: split Header component into smaller components
