+++ apps/frontend/app/components/shared/header/user-menu.tsx
@@ -1,0 +1,44 @@
+import {
+  History,
+  List,
+	Settings,
+	User,
+} from "lucide-react";
+
+import { Button } from "@/app/components/ui/button";
+import {
+	DropdownMenu,
+	DropdownMenuContent,
+	DropdownMenuItem,
+	DropdownMenuTrigger,
+} from "@/app/components/ui/dropdown-menu";
+
+export const UserMenu = () => {
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="ghost" size="icon">
+          <User className="!h-6 !w-6 transition-transform group-hover:scale-110" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent className="w-48">
+        <DropdownMenuItem>
+          <User className="mr-2 h-4 w-4" />
+          Profile
+        </DropdownMenuItem>
+        <DropdownMenuItem>
+          <List className="mr-2 h-4 w-4" />
+          My Listings
+        </DropdownMenuItem>
+        <DropdownMenuItem>
+          <History className="mr-2 h-4 w-4" />
+          Transaction History
+        </DropdownMenuItem>
+        <DropdownMenuItem>
+          <Settings className="mr-2 h-4 w-4" />
+          Settings
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  )
+}
